[API][Admin] Enable contract tests for Product Association resource
